requirements.txt
absl-py==2.3.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
docutils==0.21.2
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.1
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.2.1
protobuf==5.29.5
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
ttach==0.0.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
